move event types and properties to Constants
