+++ android/src/main/java/com/amplitude/android/internal/fragments/AutocaptureFragmentLifecycleCallbacks.kt
@@ -5,8 +5,8 @@ import androidx.annotation.RequiresApi
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import com.amplitude.android.utilities.DefaultEventUtils.Companion.screenName
-import com.amplitude.android.utilities.DefaultEventUtils.EventProperties
-import com.amplitude.android.utilities.DefaultEventUtils.EventTypes
+import com.amplitude.core.Constants.EventProperties
+import com.amplitude.core.Constants.EventTypes
 import com.amplitude.common.Logger
 
 @RequiresApi(Build.VERSION_CODES.LOLLIPOP)
