+++ android/src/main/java/com/amplitude/android/internal/gestures/AutocaptureGestureListener.kt
@@ -8,15 +8,15 @@ import com.amplitude.android.internal.ViewHierarchyScanner.findTarget
 import com.amplitude.android.internal.ViewTarget
 import com.amplitude.android.internal.locators.ViewTargetLocator
 import com.amplitude.android.utilities.DefaultEventUtils.Companion.screenName
-import com.amplitude.android.utilities.DefaultEventUtils.EventProperties.ACTION
-import com.amplitude.android.utilities.DefaultEventUtils.EventProperties.HIERARCHY
-import com.amplitude.android.utilities.DefaultEventUtils.EventProperties.SCREEN_NAME
-import com.amplitude.android.utilities.DefaultEventUtils.EventProperties.TARGET_CLASS
-import com.amplitude.android.utilities.DefaultEventUtils.EventProperties.TARGET_RESOURCE
-import com.amplitude.android.utilities.DefaultEventUtils.EventProperties.TARGET_SOURCE
-import com.amplitude.android.utilities.DefaultEventUtils.EventProperties.TARGET_TAG
-import com.amplitude.android.utilities.DefaultEventUtils.EventProperties.TARGET_TEXT
-import com.amplitude.android.utilities.DefaultEventUtils.EventTypes.ELEMENT_INTERACTED
+import com.amplitude.core.Constants.EventProperties.ACTION
+import com.amplitude.core.Constants.EventProperties.HIERARCHY
+import com.amplitude.core.Constants.EventProperties.SCREEN_NAME
+import com.amplitude.core.Constants.EventProperties.TARGET_CLASS
+import com.amplitude.core.Constants.EventProperties.TARGET_RESOURCE
+import com.amplitude.core.Constants.EventProperties.TARGET_SOURCE
+import com.amplitude.core.Constants.EventProperties.TARGET_TAG
+import com.amplitude.core.Constants.EventProperties.TARGET_TEXT
+import com.amplitude.core.Constants.EventTypes.ELEMENT_INTERACTED
 import com.amplitude.common.Logger
 import java.lang.ref.WeakReference
